Function sentry_actix::capture_actix_error

pub fn capture_actix_error(err: &Error) -> Uuid

Utility function that takes an actix error and reports it to the default hub.

This is typically not very since the actix hub is likely never bound as the default hub. It's generally recommended to use the ActixWebHubExt trait's extension method on the hub instead.